+++ refinedstorage-autocrafting-api/src/main/java/com/refinedmods/refinedstorage/api/autocrafting/task/TaskImpl.java
@@ -144,7 +144,7 @@ public class TaskImpl implements Task {
 
     @Override
     public TaskStatus getStatus() {
-        final TaskStatusBuilder builder = new TaskStatusBuilder(id, resource, amount, startTime);
+        final TaskStatusBuilder builder = new TaskStatusBuilder(id, state, resource, amount, startTime);
         double totalWeightedCompleted = 0;
         double totalWeight = 0;
         for (final AbstractTaskPattern pattern : patterns.values()) {