+++ refinedstorage-autocrafting-api/src/test/java/com/refinedmods/refinedstorage/api/autocrafting/task/TaskImplTest.java
@@ -1033,7 +1033,7 @@ class TaskImplTest {
         task.step(storage, EMPTY_SINK_PROVIDER, StepBehavior.DEFAULT, TaskListener.EMPTY);
         assertThat(task.getState()).isEqualTo(TaskState.RUNNING);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), CRAFTING_TABLE, 2, 0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.RUNNING, CRAFTING_TABLE, 2, 0)
                 .crafting(CRAFTING_TABLE, 2)
                 .crafting(OAK_PLANKS, 4)
                 .stored(OAK_PLANKS, 4)
@@ -1069,7 +1069,7 @@ class TaskImplTest {
         task.step(storage, sinkProvider, StepBehavior.DEFAULT, TaskListener.EMPTY);
         assertThat(task.getState()).isEqualTo(TaskState.RUNNING);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), IRON_PICKAXE, 1, 0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.RUNNING, IRON_PICKAXE, 1, 0)
                 .crafting(IRON_PICKAXE, 1)
                 .scheduled(IRON_INGOT, 2)
                 .processing(IRON_ORE, 1, sinkKey(IRON_INGOT_PATTERN))
@@ -1090,7 +1090,7 @@ class TaskImplTest {
         task.step(storage, sinkProvider, StepBehavior.DEFAULT, TaskListener.EMPTY);
         assertThat(task.getState()).isEqualTo(TaskState.RUNNING);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), IRON_PICKAXE, 1, 0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.RUNNING, IRON_PICKAXE, 1, 0)
                 .crafting(IRON_PICKAXE, 1)
                 .scheduled(IRON_INGOT, 1)
                 .processing(IRON_ORE, 2, sinkKey(IRON_INGOT_PATTERN))
@@ -1112,7 +1112,7 @@ class TaskImplTest {
         task.step(storage, sinkProvider, StepBehavior.DEFAULT, TaskListener.EMPTY);
         assertThat(task.getState()).isEqualTo(TaskState.RUNNING);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), IRON_PICKAXE, 1, 0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.RUNNING, IRON_PICKAXE, 1, 0)
                 .crafting(IRON_PICKAXE, 1)
                 .processing(IRON_ORE, 3, sinkKey(IRON_INGOT_PATTERN))
                 .stored(OAK_PLANKS, 2)
@@ -1131,7 +1131,7 @@ class TaskImplTest {
         storage.insert(IRON_INGOT, 2, Action.EXECUTE, Actor.EMPTY);
         assertThat(task.getState()).isEqualTo(TaskState.RUNNING);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), IRON_PICKAXE, 1, 0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.RUNNING, IRON_PICKAXE, 1, 0)
                 .crafting(IRON_PICKAXE, 1)
                 .processing(IRON_ORE, 1, sinkKey(IRON_INGOT_PATTERN))
                 .stored(OAK_PLANKS, 2)
@@ -1152,7 +1152,7 @@ class TaskImplTest {
         storage.insert(IRON_INGOT, 1, Action.EXECUTE, Actor.EMPTY);
         assertThat(task.getState()).isEqualTo(TaskState.RUNNING);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), IRON_PICKAXE, 1, 0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.RUNNING, IRON_PICKAXE, 1, 0)
                 .crafting(IRON_PICKAXE, 1)
                 .stored(OAK_PLANKS, 2)
                 .stored(STICKS, 4)
@@ -1176,7 +1176,7 @@ class TaskImplTest {
         task.step(storage, sinkProvider, StepBehavior.DEFAULT, TaskListener.EMPTY);
         assertThat(task.getState()).isEqualTo(TaskState.RETURNING_INTERNAL_STORAGE);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), IRON_PICKAXE, 1, 0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.RETURNING_INTERNAL_STORAGE, IRON_PICKAXE, 1, 0)
                 .stored(OAK_PLANKS, 2)
                 .stored(STICKS, 2)
                 .build(1.0));
@@ -1198,7 +1198,7 @@ class TaskImplTest {
         task.step(storage, sinkProvider, StepBehavior.DEFAULT, TaskListener.EMPTY);
         assertThat(task.getState()).isEqualTo(TaskState.COMPLETED);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), IRON_PICKAXE, 1, 0).build(1.0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.COMPLETED, IRON_PICKAXE, 1, 0).build(1.0)
         );
         assertThat(copyInternalStorage(task)).isEmpty();
         assertThat(ironOreSink.getAll()).usingRecursiveFieldByFieldElementComparator().containsExactly(
@@ -1225,7 +1225,7 @@ class TaskImplTest {
         // Act & assert
         task.step(storage, sinkProvider, StepBehavior.DEFAULT, TaskListener.EMPTY);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), IRON_INGOT, 2, 0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.RUNNING, IRON_INGOT, 2, 0)
                 .scheduled(IRON_INGOT, 1)
                 .stored(IRON_ORE, 1)
                 .processing(IRON_ORE, 1, sinkKey(IRON_INGOT_PATTERN))
@@ -1234,7 +1234,7 @@ class TaskImplTest {
         sinkProvider.put(IRON_INGOT_PATTERN, ExternalPatternSink.Result.REJECTED);
         task.step(storage, sinkProvider, StepBehavior.DEFAULT, TaskListener.EMPTY);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), IRON_INGOT, 2, 0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.RUNNING, IRON_INGOT, 2, 0)
                 .scheduled(IRON_INGOT, 1)
                 .stored(IRON_ORE, 1)
                 .processing(IRON_ORE, 1, sinkKey(IRON_INGOT_PATTERN))
@@ -1254,7 +1254,7 @@ class TaskImplTest {
         // Act & assert
         task.step(storage, sinkProvider, StepBehavior.DEFAULT, TaskListener.EMPTY);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), IRON_INGOT, 2, 0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.RUNNING, IRON_INGOT, 2, 0)
                 .scheduled(IRON_INGOT, 1)
                 .stored(IRON_ORE, 1)
                 .processing(IRON_ORE, 1, sinkKey(IRON_INGOT_PATTERN))
@@ -1263,7 +1263,7 @@ class TaskImplTest {
         sinkProvider.remove(IRON_INGOT_PATTERN);
         task.step(storage, EMPTY_SINK_PROVIDER, StepBehavior.DEFAULT, TaskListener.EMPTY);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), IRON_INGOT, 2, 0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.RUNNING, IRON_INGOT, 2, 0)
                 .scheduled(IRON_INGOT, 1)
                 .stored(IRON_ORE, 1)
                 .processing(IRON_ORE, 1, sinkKey(IRON_INGOT_PATTERN))
@@ -1283,7 +1283,7 @@ class TaskImplTest {
         // Act & assert
         task.step(storage, sinkProvider, StepBehavior.DEFAULT, TaskListener.EMPTY);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), IRON_INGOT, 2, 0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.RUNNING, IRON_INGOT, 2, 0)
                 .scheduled(IRON_INGOT, 1)
                 .stored(IRON_ORE, 1)
                 .processing(IRON_ORE, 1, sinkKey(IRON_INGOT_PATTERN))
@@ -1292,7 +1292,7 @@ class TaskImplTest {
         sinkProvider.put(IRON_INGOT_PATTERN, ExternalPatternSink.Result.LOCKED);
         task.step(storage, sinkProvider, StepBehavior.DEFAULT, TaskListener.EMPTY);
         assertThat(task.getStatus()).usingRecursiveComparison(STATUS_CONFIG).isEqualTo(
-            new TestTaskStatusBuilder(task.getId(), IRON_INGOT, 2, 0)
+            new TestTaskStatusBuilder(task.getId(), TaskState.RUNNING, IRON_INGOT, 2, 0)
                 .scheduled(IRON_INGOT, 1)
                 .stored(IRON_ORE, 1)
                 .processing(IRON_ORE, 1, sinkKey(IRON_INGOT_PATTERN))