+++ refinedstorage-autocrafting-api/src/test/java/com/refinedmods/refinedstorage/api/autocrafting/task/TestTaskStatusBuilder.java
@@ -5,6 +5,7 @@ import com.refinedmods.refinedstorage.api.autocrafting.status.TaskStatusBuilder;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 
 import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Map;
 import javax.annotation.Nullable;
 
@@ -15,10 +16,13 @@ import javax.annotation.Nullable;
  */
 public class TestTaskStatusBuilder {
     private final TaskStatus.TaskInfo info;
+    private final TaskState state;
     private final Map<ResourceKey, MutableItem> items = new LinkedHashMap<>();
 
-    public TestTaskStatusBuilder(final TaskId id, final ResourceKey resource, final long amount, final long startTime) {
+    public TestTaskStatusBuilder(final TaskId id, final TaskState state,
+                                 final ResourceKey resource, final long amount, final long startTime) {
         this.info = new TaskStatus.TaskInfo(id, resource, amount, startTime);
+        this.state = state;
     }
 
     public TestTaskStatusBuilder stored(final ResourceKey resource, final long stored) {
@@ -64,7 +68,7 @@ public class TestTaskStatusBuilder {
     }
 
     public TaskStatus build(final double percentageCompleted) {
-        return new TaskStatus(info, percentageCompleted, items.entrySet().stream().map(entry -> new TaskStatus.Item(
+        final List<TaskStatus.Item> mappedItems = items.entrySet().stream().map(entry -> new TaskStatus.Item(
             entry.getKey(),
             entry.getValue().type,
             entry.getValue().sinkKey,
@@ -72,7 +76,8 @@ public class TestTaskStatusBuilder {
             entry.getValue().processing,
             entry.getValue().scheduled,
             entry.getValue().crafting
-        )).toList());
+        )).toList();
+        return new TaskStatus(info, state, percentageCompleted, mappedItems);
     }
 
     private static class MutableItem {